Fixed a MSVC warning involving int64 conversion to double

+++ client/ai.cpp
@@ -4572,7 +4572,7 @@ std::uint8_t ai::strategyRandom(ai::Board board)
 
     std::random_device rd;
     std::default_random_engine engine(rd());
-    std::uniform_real_distribution<double> dist(0, openCells.size());
+    std::uniform_real_distribution<double> dist(0.0, static_cast<double>(openCells.size()));
 
     std::uint8_t choice = static_cast<std::uint8_t>(dist(engine));
     return choice;
